[Fix] 스터디, 스터디 피드백 api 전체 조회 기능 추가 및 전체적으로 코드 로직에 맞게 수정

+++ src/apis/study.js
@@ -10,8 +10,6 @@ const studyApi = {
     try {
       // 스터디 개설
       const user_id = req.user._id;
-      console.log(user_id);
-      res.end();
       const { study_name, title, content, deadline, headcount, chat_link, status } = req.body;
 
       const createInfo = {
@@ -53,9 +51,6 @@ const studyApi = {
   async applyStudy(req, res, next) {
     try {
       const user_id = req.user._id;
-      // const studyId = req.query;
-      // const study_id = mongoose.Types.ObjectId(studyId);
-      // console.log(typeof study_id);
       const { study_id, goal } = req.body;
       const createInfo = {
         user_id,
@@ -79,11 +74,13 @@ const studyApi = {
   /**스터디 신청 수락*/
   async acceptStudy(req, res, next) {
     try {
+      // 스터디장 권한 판단
       const { user_id, study_id } = req.params;
       // const member = StudyRelation.findOne({ user_id: member_id });
       // if (!member || member.is_leader === 0)
       //   throw new Error('You do not have authorization to accept study applications.');
 
+      // 스터디원 권한 판단
       // const leader_id = req.user._id;
       // const leader = StudyRelation.findOne({ user_id: leader_id });
       // if (!leader || leader.is_leader === 1) throw new Error('Only leader can accept');
@@ -106,13 +103,27 @@ const studyApi = {
     }
   },
 
-  /**스터디 정보 조회*/
+  /**스터디 정보 조회(전체)*/
   async getStudy(req, res, next) {
+    try {
+      const foundStudy = await Study.find({});
+      if (!foundStudy) throw new Error('Not found');
+      res.status(200).json(foundStudy);
+    } catch (error) {
+      console.log(error);
+      res.status(426).json({
+        code: 426,
+        message: 'wrong request',
+      });
+    }
+  },
+
+  /**스터디 정보 조회(개별)*/
+  async getStudyOne(req, res, next) {
     try {
       const { study_id } = req.params;
       const foundStudy = await Study.findOne({ _id: study_id });
       if (!foundStudy) throw new Error('Not found');
-
       res.status(200).json(foundStudy);
     } catch (error) {
       console.log(error);
@@ -126,8 +137,14 @@ const studyApi = {
   /**스터디 정보 수정*/
   async updateStudy(req, res, next) {
     try {
+      // 스터디장 권한 판단
+      const leader_id = req.user._id;
+      const foundRelation = await StudyRelation.findOne({ user_id: leader_id });
+      if (!foundRelation) throw new Error('Not found');
+      if (foundRelation.is_leader != 1) throw new Error('Not authorization');
+
+      // 스터디 정보 수정
       const { study_id } = req.params;
-      // 스터디장인지 판단
       const { study_name, title, content, deadline, headcount, chat_link, status } = req.body;
       const updateInfo = {
         study_name,
@@ -149,16 +166,23 @@ const studyApi = {
     }
   },
 
-  /**스터디 회원 관리*/ // 그 유저가 쓴 피드백도 모두 삭제 해야함
+  /**스터디 회원 관리*/
   async deleteUser(req, res, next) {
     try {
-      const { user_id } = req.user._id;
-      const foundRelation = await StudyRelation.findOne({ user_id });
+      // 스터디장 권한 판단
+      const leader_id = req.user._id;
+      const foundRelation = await StudyRelation.findOne({ user_id: leader_id });
       if (!foundRelation) throw new Error('Not found');
+      if (foundRelation.is_leader != 1) throw new Error('Not authorization');
 
-      const deletedRelation = await StudyRelation.deleteOne({ user_id }); // 특정 유저 스터디에서 삭제
+      // 스터디 회원 관리
+      const member_id = req.params;
+      const deletedRelation = await StudyRelation.deleteOne({ user_id: member_id }); // 특정 유저 스터디에서 삭제
       res.study_relation = deletedRelation;
       res.status(200).json(deletedRelation);
+
+      const deletedFeedback = await StudyFeedback.deleteOne({ user_id: member_id }); // 특정 유저가 쓴 피드백을 삭제
+      res.study_feedback = deletedFeedback;
     } catch (error) {
       console.log(error);
       res.status(426).json({
@@ -167,17 +191,29 @@ const studyApi = {
       });
     }
   },
+
   /**스터디 삭제*/
   async deleteStudy(req, res, next) {
     try {
-      const { study_id } = req.user._id;
+      // 스터디장 권한 판단
+      const leader_id = req.user._id;
+      const foundRelation = await StudyRelation.findOne({ user_id: leader_id });
+      if (!foundRelation) throw new Error('Not found');
+      if (foundRelation.is_leader != 1) throw new Error('Not authorization');
+
+      // 스터디 삭제
+      const { study_id } = req.params;
       const foundStudy = await Study.findOne({ _id: study_id });
       if (!foundStudy) throw new Error('Not found');
-
       const deletedStudy = await Study.deleteOne({ _id: study_id }); // 스터디 삭제
+      res.study = deletedStudy;
+      res.status(200).json(deletedStudy);
+
       const deletedRelation = await StudyRelation.deleteMany({ study_id }); // 해당 스터디 아이디 관계 모두 삭제
       res.study_relation = deletedRelation;
-      res.status(200).json(deletedStudy);
+
+      const deletedFeedback = await StudyFeedback.deleteMany({ study_id }); // 해당 스터디 피드백, 댓글 모두 삭제
+      res.study_feedback = deletedFeedback;
     } catch (error) {
       console.log(error);
       res.status(426).json({
